Compact kind
