+++ crates/maple_cli/src/dumb_analyzer/find_usages/search_engine/regex/worker.rs
@@ -4,10 +4,9 @@ use std::path::PathBuf;
 use anyhow::Result;
 use rayon::prelude::*;
 
+use super::definition::{build_full_regexp, get_comments_by_ext, is_comment, DefinitionKind};
 use crate::process::AsyncCommand;
 use crate::tools::ripgrep::{Match, Word};
-
-use super::definition::{build_full_regexp, get_comments_by_ext, is_comment, DefinitionKind};
 
 /// Executes `command` as a child process.
 ///
